move model and generate into tsr __init__
